updated functions location and deleted duplicated functions
