+++ SnowBear/functions/bbui.py
@@ -38,3 +38,43 @@ def load_page_navigation_data(page_navigation_data):
         if service['sub_navigation']:
             page_navigation_data[0]['sub_navigation_elements'] = load_yaml(service['path'].replace('.','/').replace('main','navigation.yml'))
     return page_navigation_data
+
+def update_yaml(yamlFile,yamlPath,data):
+  """This function inserts yaml data into a file
+  
+  Args:
+      yamlFile ([string]): path to the file to write to
+      yamlPath ([String]): list of the path to get to the attribute
+      data ([string]): data that you want to insert into the path in the file. 
+
+      for example, if your file looks like this :       
+        general_settings:
+          root_path: 
+
+      then insertFile("file.yml",['general_settings','root_path'],test)      
+      will result in your file looking like this:
+      general_settings:
+          root_path: test
+
+  """
+  if(len(yamlPath)==2):
+    stream=open(yamlFile,'r')
+    fileData=yaml.safe_load(stream)
+    fileData[yamlPath[0]][yamlPath[1]]=data
+    with open(yamlFile, 'w', encoding='utf8') as outfile:
+      outfile.write(yaml.dump(fileData,default_flow_style=False))
+  elif(len(yamlPath)==3):
+    stream=open(yamlFile,'r')
+    fileData=yaml.safe_load(stream)
+    fileData[yamlPath[0]][yamlPath[1]][yamlPath[2]]=data
+    with open(yamlFile, 'w', encoding='utf8') as outfile:
+      outfile.write(yaml.dump(fileData,default_flow_style=False))
+  elif(len(yamlPath)==4):
+    stream=open(yamlFile,'r')
+    fileData=yaml.safe_load(stream)
+    fileData[yamlPath[0]][yamlPath[1]][yamlPath[2]][yamlPath[3]]=data
+    with open(yamlFile, 'w', encoding='utf8') as outfile:
+      outfile.write(yaml.dump(fileData,default_flow_style=False))
+  else:
+    print("Path superior to 4 is not supported")
+
